# Faster 'uniform' and 'uniformR' for Bool