Benchmark:

```
How many random numbers can we generate in a second on one thread?
  First, timing System.Random.next:                             BEFORE                 AFTER
  Second, timing System.Random.random at different types:
      1,859,311 randoms generated [System.Random Bools]       ~ 932 cycles/int       ~ 22.32 cycles/int
  Next timing range-restricted System.Random.randomR:
      1,608,582 randoms generated [System.Random Bools]       ~ 1,077 cycles/int     ~ 22.46 cycles/int
Now 8 threads, reporting mean randoms-per-second-per-thread:
  First, timing System.Random.next:
  Second, timing System.Random.random at different types:
        726,606 randoms generated [System.Random Bools]       ~ 2,385 cycles/int     ~ 43.49 cycles/int
  Next timing range-restricted System.Random.randomR:
        687,065 randoms generated [System.Random Bools]       ~ 2,523 cycles/int     ~ 46.32 cycles/int
```